Refactoring of code
Refactored code so that type hints are correct and common Python standards are followed throughout the majority of the program.

Decided not to touch the Webhook data portion of the code, as making this section abide to common standards would be a significant amount of effort for little return.

Fixed a few typos as well.